% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_isbn}
\alias{check_isbn}
\title{Verify that all ISBN numbers are unique}
\usage{
check_isbn(data, data_source = c("Google", "Zotero"))
}
\arguments{
\item{data}{Dataframe including at least one column called "isbn"}

\item{data_source}{Either "Google" or "Zotero"}
}
\value{
TRUE if there are no duplicate ISBN numbers or an error stating
which ISBN numbers are duplicates.
}
\description{
Verify that all ISBN numbers are unique
}
