% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_zotero}
\alias{get_zotero}
\title{Import data from Zotero}
\usage{
get_zotero(groupID, api_key)
}
\arguments{
\item{groupID}{Zotero groupID (not group name);
precedes group name in group URL}

\item{api_key}{API key issued by the group admin}
}
\value{
A tibble
}
\description{
Download a bibliography from a Zotero group
and convert it to a tibble (tidy dataframe).
}
\details{
To find the groupID of your private Zotero group, look at its URL.
The groupID is the part just before the group name. For example,
the groupID of

https://www.zotero.org/groups/2220437/mygroup_01/

is 2220437.
}
