% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initiate_booklist}
\alias{initiate_booklist}
\title{Create a google sheet with data imported from Zotero}
\usage{
initiate_booklist(sheet_name, groupID, api_key)
}
\arguments{
\item{sheet_name}{Name to use for the google sheet}

\item{groupID}{ID number of Zotero group with bibliography}

\item{api_key}{API key needed to access Zotero group}
}
\value{
A \code{\link[googlesheets]{googlesheets}} object; a google sheet
will be created using data from Zotero, and new columns for purchase and
loan status will be added.
}
\description{
The google sheet will be created in the user's google drive root folder.
}
