% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_gs_data}
\alias{load_gs_data}
\title{Load a google sheet as a dataframe}
\usage{
load_gs_data(table)
}
\arguments{
\item{table}{Name of google sheet}
}
\value{
dataframe
}
\description{
From https://shiny.rstudio.com/articles/persistent-data-storage.html
}
