% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_booklist}
\alias{update_booklist}
\title{Synchronize a Zotero bibliography and a Google Sheet}
\usage{
update_booklist(groupID, api_key, sheet_name)
}
\arguments{
\item{groupID}{ID number of Zotero group with bibliography}

\item{api_key}{API key needed to access Zotero group}

\item{sheet_name}{Name of the google sheet}
}
\value{
A \code{\link[googlesheets]{googlesheets}} object;
the google sheet should be updated with new book entries from Zotero, if any.
}
\description{
It is assumed that the google sheet is located in the root folder of the
user's Google Drive.
}
